+++ internal/app/models/servers.go
@@ -1,10 +1,10 @@
 package models
 
 type Server struct {
-	ID           int     `gorm:"primaryKey;autoIncrement"`
+	ID           uint    `gorm:"primaryKey;autoIncrement"`
 	IP           string  `gorm:"size:15;unique;not null"`
-	CountryID    int     `gorm:"not null"`
+	CountryID    uint    `gorm:"not null"`
 	Country      Country `gorm:"foreignKey:CountryID;references:ID"`
-	ChannelSpeed int     `gorm:"not null"`
-	Port         int     `gorm:"not null"`
+	ChannelSpeed uint64  `gorm:"not null"`
+	Port         uint    `gorm:"not null"`
 }
